Struct linera_alloy_sol_types::abi::Decoder

pub struct Decoder<'de> { /* private fields */ }

The Decoder wraps a byte slice with necessary info to progressively deserialize the bytes into a sequence of tokens.

Usage Note

While the Decoder contains the necessary info, the actual deserialization is done in the crate::SolType trait.

Implementations

impl<'de> Decoder<'de>

pub const fn new(buf: &'de [u8], validate: bool) -> Self

Instantiate a new decoder from a byte slice and a validation flag.

If validate is true, the decoder will check that the bytes conform to expected type limitations, and that the decoded values can be re-encoded to an identical bytestring.

pub const fn offset(&self) -> usize

Returns the current offset in the buffer.

pub const fn remaining(&self) -> Option<usize>

Returns the number of bytes in the remaining buffer.

pub const fn remaining_words(&self) -> usize

Returns the number of words in the remaining buffer.

pub fn remaining_buf(&self) -> Option<&'de [u8]>

Returns a reference to the remaining bytes in the buffer.

pub const fn is_empty(&self) -> bool

Returns whether the remaining buffer is empty.

pub const fn validate(&self) -> bool

Returns true if this decoder is validating type correctness.

pub fn set_validate(&mut self, validate: bool)

Set whether to validate type correctness.

pub fn raw_child(&self) -> Result<Self>

Create a child decoder, starting at offset bytes from the current decoderโs offset.

See child.

pub fn child(&self, offset: usize) -> Result<Decoder<'de>, Error>

Create a child decoder, starting at offset bytes from the current decoderโs offset. The child decoder shares the buffer and validation flag.

pub fn peek<I: SliceIndex<[u8]>>( &self, index: I, ) -> Result<&'de I::Output, Error>

Peek into the buffer.

pub fn peek_len_at(&self, offset: usize, len: usize) -> Result<&'de [u8], Error>

Peek a slice of size len from the buffer at a specific offset, without advancing the offset.

pub fn peek_len(&self, len: usize) -> Result<&'de [u8], Error>

Peek a slice of size len from the buffer without advancing the offset.

pub fn peek_word_at(&self, offset: usize) -> Result<&'de Word, Error>

Peek a word from the buffer at a specific offset, without advancing the offset.

pub fn peek_word(&self) -> Result<&'de Word, Error>

Peek the next word from the buffer without advancing the offset.

pub fn peek_offset_at(&self, offset: usize) -> Result<usize>

Peek a usize from the buffer at a specific offset, without advancing the offset.

pub fn peek_offset(&self) -> Result<usize>

Peek a usize from the buffer, without advancing the offset.

pub fn take_word(&mut self) -> Result<&'de Word, Error>

Take a word from the buffer, advancing the offset.

pub fn take_indirection(&mut self) -> Result<Decoder<'de>, Error>

Return a child decoder by consuming a word, interpreting it as a pointer, and following it.

pub fn take_offset(&mut self) -> Result<usize>

Takes a usize offset from the buffer by consuming a word.

pub fn take_slice(&mut self, len: usize) -> Result<&'de [u8]>

Takes a slice of bytes of the given length by consuming up to the next word boundary.

pub fn take_slice_unchecked(&mut self, len: usize) -> Result<&'de [u8]>

Takes a slice of bytes of the given length.

pub fn take_offset_from(&mut self, child: &Decoder<'de>)

Takes the offset from the child decoder and sets it as the current offset.

pub fn set_offset(&mut self, offset: usize)

Sets the current offset in the buffer.

pub fn decode<T: Token<'de>>(&mut self) -> Result<T>

Decodes a single token from the underlying buffer.

pub fn decode_sequence<T: Token<'de> + TokenSeq<'de>>(&mut self) -> Result<T>

Decodes a sequence of tokens from the underlying buffer.