Struct linera_alloy_sol_types::abi::token::PackedSeqToken

pub struct PackedSeqToken<'a>(pub &'a [u8]);

A Packed Sequence - bytes or string

Tuple Fields
0: &'a [u8]
Implementations

impl PackedSeqToken<'_>

pub fn into_vec(self) -> Vec<u8>

Instantiate a new Vec by copying the underlying slice.

pub fn into_bytes(self) -> Bytes

Instantiate a new Bytes by copying the underlying slice.

pub const fn as_slice(&self) -> &[u8]

Returns a reference to the slice.